+++ jest.config.js
@@ -1,0 +1,7 @@
+/* eslint-disable max-len */
+// For a detailed explanation regarding each configuration property, visit:
+// https://jestjs.io/docs/en/configuration.html
+
+module.exports = {
+  setupFiles: ['<rootDir>/enzyme.config.js'],
+}