+++ CHANGES.md
@@ -1,9 +1,12 @@
 ## 0.0.2
 > Jun 23, 2019
 
-- Added tests using `jest` for:
+- Added tests using `Jest` for:
   - Helper functions: _getNestedValue_, _getLastValue_
 - Added the example distribution file for dynamic demo deployment
+- Added a _keyId_ prop to let the rendering use it as React's _key_ in loops
+- Added tests using `Enzyme` for:
+  - Helper component: _Link_
 
 ## 0.0.1
 > Jun 19, 2019